pyproject.toml
[tool.poetry]
name = "service-spms"
version = "0.1.0"
description = ""
authors = ["idabblewith <109533655+idabblewith@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "<3.15,>=3.12.11"
beautifulsoup4 = "^4.12.2"
brotli = "^1.1.0"
django-cors-headers = "^4.3.0"
django-environ = "^0.12.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
requests = "^2.31.0"
whitenoise = { extras = ["brotli"], version = "^6.6.0" }
gunicorn = ">=22.0.0"
pandas = "^3.0.0"
sqlparse = ">=0.5.0"
chardet = "^5.2.0"
cssutils = "^2.9.0"
tqdm = "^4.66.2"
sentry-sdk = "^2.1.1"
psycopg = "^3.2.2"
psycopg-binary = "^3.2.2"
psycopg-pool = "^3.2.3"
bleach = "^6.2.0"
tinycss2 = "^1.4.0"
html2text = "^2025.4.15"
dj-database-url = "^3.0.0"
dbca-utils = "^2.1.3"
pillow = "^12.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-django = "^4.8.0"
pytest-cov = "^7.0.0"
pytest-xdist = "^3.5.0"
factory-boy = "^3.3.0"
faker = "^40.0.0"
pytest-mock = "^3.12.0"
pytest-split = "^0.11.0"
pre-commit = "^4.0.0"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
bandit = "^1.8.0"
autoflake = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/venv/*", "*/.venv/*"]

# Bandit configuration
[tool.bandit]
exclude_dirs = [
    "*/tests/*",
    "*/test_*.py",
    "*/migrations/*",
    "*/.venv/*",
    "*/venv/*",
    "_refactor_archive"  # Exclude old archived code
]
# Skip low-priority warnings and false positives:
# B101: assert_used (common in Django/tests)
# B601: shell_injection (parameterized shell calls are safe)
# B404: subprocess import (legitimate use for PDF generation)
# B607: partial executable path (prince PDF generator is trusted)
# B603: subprocess without shell (safe usage without shell=True)
# B105: hardcoded password (false positive for rate limit strings like '3/hour')
skips = ["B101", "B601", "B404", "B607", "B603", "B105"]
